feat: add collection management components
