+++ apps/web/src/pages/Settings/index.tsx
@@ -1,0 +1,3 @@
+// Назначение: экспорт страницы управления коллекциями
+// Основные модули: CollectionsPage
+export { default } from "./CollectionsPage";
